Fixed some bugs on the examples.

+++ examples/vector/vector.c
@@ -49,7 +49,7 @@ void vector_delete(vector *v, size_t index) {
     v->items[index] = NULL;
 
     /* Reset the rest of the elements forwards */
-    for(int i = index; i < v->len - 1; i++) {
+    for(size_t i = index; i < v->len - 1; i++) {
         v->items[i] = v->items[i + 1];
         v->items[i + 1] = NULL;
     }
